requirements.txt
pypdf>=4.0.0
Pillow>=10.0.0
reportlab>=4.0.0
python-docx>=1.1.0
docx2pdf>=0.1.8
extract-msg>=0.48.0
click>=8.0.0
